Added conf to app, reload to Dockerfile
The reload should probably not be on production. To help this, it might work
to have two separate dockerfiles, dev and prod. The prod one will have the
regular go command and not get the packages. Another way would be to have the
docker run command on prod pass in a different entrypoint or cmd.

TODO:
- Change Dockerfile to correctly use ENTRYPOINT and CMD together
- App file is a mess